Aggiornamenti a AttachmentController e AttachmentLogic
